pyproject.toml
[project]
name = "openbb-hka"
version = "0.3.1"
description = "Dashboard for your financial data in China and Hong Kong market"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "nbformat>=5.10.4",
    "openbb>=4.1.3",
    "openbb-akshare>=0.9.23",
    "openbb-tushare>=0.3.5",
    "openbb-core==1.4.8",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.34.2",
    "openbb-ai>=1.7.5",
    "sse-starlette>=3.0.2",
    "openai>=1.109.1",
    "magentic>=0.40.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "openbb-devtools>=1.4.4",
    "pytest>=7.4.4",
]
